requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2023.10
cryptography==41.0.7
defusedxml==0.7.1
distlib==0.3.7
Django==4.2.4
django-adaptors==0.2.5
django-allauth==0.58.2
django-crispy-forms==2.1
django-extensions==3.2.3
django-otp==1.3.0
django-shortuuidfield==0.1.3
django-timezone-field==6.1.0
filelock==3.13.1
idna==3.6
numpy==1.26.2
oauthlib==3.2.2
paypal-checkout-serversdk==1.0.3
paypalhttp==1.0.1
Pillow==10.1.0
platformdirs==4.1.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.3.0
pyotp==2.9.0
pyparsing==3.1.1
python-dotenv==1.0.1
python3-openid==3.2.0
reportlab==4.0.7
requests==2.31.0
requests-oauthlib==1.3.1
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
virtualenv==20.25.0
whitenoise==6.8.2


